added task 1029 from Foncode

+++ Foncode1032.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 
-//*
+/*
 int main() {
-	int in; std::cin >> in;
+	long in; std::cin >> in;
 	std::cout << in * (in - 1) * 2 + 1;
 }
 //*/
